requirements.txt
sqlalchemy
fastapi[standard]
uvicorn
pymysql
python-dotenv
pydantic
google-generativeai